fix type

+++ epochX/sycl/gg_ttgg.auto/SubProcesses/P1_Sigma_sm_gg_ttxgg/CPPProcess.cc
@@ -51,7 +51,7 @@ namespace Proc
 #endif
 
   // Helicity combinations (and filtering of "good" helicity combinations)
-  static int cHel[ncomb][npar];
+  static short cHel[ncomb][npar];
 
   //--------------------------------------------------------------------------
